fix(admin): use GetInt64 for user_id when creating API keys
Auth middleware stores user_id in the gin context as int64, but
CreateAPIKey read it with GetInt, whose strict type assertion fails
silently against int64 and returns 0. Every created key ended up with
created_by=0, violating the api_keys_created_by_fkey constraint since
user id 0 doesn't exist.

+++ backend/handler/apikeys.go
@@ -32,7 +32,7 @@ func hashAPIKey(key string) string {
 
 func CreateAPIKey(database *sql.DB) gin.HandlerFunc {
 	return func(c *gin.Context) {
-		userID := c.GetInt("user_id")
+		userID := c.GetInt64("user_id")
 
 		var req struct {
 			Name            string             `json:"name" binding:"required"`
